+++ game/states/sm_checkpoint.h
@@ -13,6 +13,7 @@
 #include "c_light.h"
 #include "c_animation.h"
 #include "c_noise.h"
+#include "c_texture.h"
 
 #define MAKE_STRING(TYPE) [[maybe_unused]] static constexpr auto make_string(Types<TYPE>) { return Str{#TYPE}; };
 
@@ -42,8 +43,13 @@ namespace Verse::State::Checkpoint
         using Default::handle;
         
         void onEnter(const ActivateEvent &e) {
+            Component::Texture* tex = e.s->getComponent<Component::Texture>(e.eid);
+            tex->use_collider_transform = false;
+            
             Component::Collider* col = e.s->getComponent<Component::Collider>(e.eid);
             e.s->checkpoints.push_back(col->transform.pos);
+            col->transform.pos -= Vec2(24, 16);
+            col->transform.size += Vec2(48, 32);
             
             Component::Light* light = e.s->getComponent<Component::Light>(e.eid);
             light->radius = 75;
@@ -55,8 +61,6 @@ namespace Verse::State::Checkpoint
             Component::Noise* noise = e.s->getComponent<Component::Noise>(e.eid);
             if (noise != nullptr)
                 noise->enabled = true;
-            
-            e.s->removeComponent<Component::Collider>(e.eid);
         }
         
         To<DeactivatedState> handle(const DeactivateEvent &e) { return To<DeactivatedState>(); };